Create the Application Date layer
The application's Data layer was created, where the Repository and the Data Sources containing the ServiceProvider are centralized.
We are going to use the repository pattern to handle the API data.
In the repository class, we need to pass the retrofit service instance
to perform the network call. In this package were the data classes.
Retrofit classes and response classes and application endpoints.
Response classes (classes containing retrofit and annotations).
These layers do not mix with the presentation layers.
Response class (Object that represents the initial root of the structure: responsible for mapping the API Json structure)